+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - List related docs in OperationDetails and ServiceDetails view
 - Add PROCESSING requests in the dashboard of the main page
+- Display only docs that are not linked to services or operations in ListDoc
 
 ## Feature